auth/app/requirements.txt
click==7.1.2
Flask==1.1.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Werkzeug==1.0.1
mysql-connector-python==8.0.22
marshmallow==3.10.0
Flask-SQLAlchemy==2.4.4
psutil==5.8.0
PyJWT==2.0.0
